+++ chalice/app.py
@@ -12,6 +12,7 @@ import os
 from chalice import Chalice
 
 from chalicelib.auth import AuthHandler
+from chalicelib.database_handle import DatabaseHandle
 from chalicelib.template_handler import TemplateHandler
 from chalicelib.utilities import Utilities
 
@@ -128,3 +129,92 @@ app.load_views(
         'admin', 'audit',  # periodic and triggered lambdas
     ]
 )
+
+
+###
+# admin.py lambdas moved here, because they are not loaded properly from a separate file
+###
+
+
+@app.lambda_function()
+def database_create_tables(event, context):
+    try:
+        dbh = DatabaseHandle(app)
+        table_list = []
+        message = ""
+        for table_name in event['Tables']:
+            model = dbh.get_model(table_name)
+            table_list.append(model)
+        created = dbh.create_tables(table_list)
+    except Exception as err:
+        app.log.error(str(err))
+        created = False
+        message = str(err)
+    if created:
+        response = ", ".join(event['Tables'])
+    else:
+        response = f"Table create failed: {message}"
+    return response
+
+
+@app.lambda_function()
+def database_create_item(event, context):
+    try:
+        dbh = DatabaseHandle(app)
+        item = dbh.create_item(event)
+        data = item.serialize()
+        json_data = app.utilities.to_json(data)
+    except Exception as err:
+        app.log.error(str(err))
+        json_data = None
+    return json_data
+
+
+@app.lambda_function()
+def database_create_items(event, context):
+    try:
+        dbh = DatabaseHandle(app)
+        created = dbh.create_items(event)
+        json_data = app.utilities.to_json(created)
+    except Exception as err:
+        app.log.error(str(err))
+        json_data = None
+    return json_data
+
+
+@app.lambda_function()
+def database_get_item(event, context):
+    app.log.debug('database_get_item function')
+    try:
+        dbh = DatabaseHandle(app)
+        item = dbh.get_item(event)
+        data = item.serialize()
+        json_data = app.utilities.to_json(data)
+    except Exception as err:
+        app.log.error(str(err))
+        json_data = None
+    return json_data
+
+
+@app.lambda_function()
+def database_run(event, context):
+    try:
+        dbh = DatabaseHandle(app)
+        dbh.set_credentials(event['User'], event['Password'])
+        status = dbh.execute_commands(event['Commands'])
+    except Exception as err:
+        app.log.error(str(err))
+        status = False
+    return status
+
+
+@app.lambda_function()
+def database_list_models(event, context):
+    try:
+        dbh = DatabaseHandle(app)
+        models = dbh.get_models()
+        tables = models.keys()
+    except Exception as err:
+        app.log.error(str(err))
+        tables = []
+    return tables
